requirements.txt
celery==5.2.7
Django==4.2
django-ninja==0.21.0
djangorestframework==3.14.0
pandas==2.0.0
pip-chill==1.0.3
pylint==2.17.2
python-dotenv==1.0.0
redis==4.5.4